Changed the order of the ItemDeduplication arguments
This fixes the issue of removing recipients form To, CC when they are
present in Bto or BCC.

We removed the code which removed an Activity's actor from the
recipients list. I see no reason why we would override the intention of
the Activity creator which added it.

Additionally we made the Audience property immutable in this context.